Add rake task for reprocessing only meta data
The previous way of generating meta data for existing
Paperclip::Attachments was to reprocess them all.  For existing
implementations with lots of records and many styles that take time to
resize and process thumbnails that could be a lot of overkill.   This
creates a simple task that only generates the meta data for existing
attachments and their styles.